requirementss.txt
Django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn
psycopg2-binary
gunicorn
dj-database-url
whitenoise
psycopg2-binary